Fix flow control
Morph can push messages onto a queue faster than Angler can pull them
off the queue, and RabbitMQ's own back-pressure mechanisms don't seem to
work, so we have tried to implement our own flow control mechanism.
However, the first attempt didn't work.

The first attempt involved the producer working out how fast the
consumer was consuming the queue, and then sleeping between producing
each record, so as to not produce messages faster than the consumer was
consuming them.  This was stupid and I should have my licence to
computer revoked.  The problem is that the rate that messages are
consumed is either limited by a bottleneck on the consumer (in which
case what we were doing makes sense) or is limited by the rate that
messages are produced (cos it can't, y'know, consume messages that
aren't being produced).  So if the producer hits a slow patch, the
producer will see that the consumer is also consuming slowly, so it'll
subsequently limit itself so that it doesn't overwhelm the consumer,
which is actually fine and not at all overwhelmed.

This commit should fix that.  Every 1000 records, the producer checks to
see whether the bot_record_consumer queue has grown above a threshold.
If the queue is too big, it sleeps for a random amount of time, before
checking again.  There may be several other producers, so the random
length of the sleep is to allow each producer to get a chance at being
able to continue producing once the queue size drops below the
threshold.